Support store up to 2GB on IIS and 4GB on Kestrel

+++ src/Microsoft.Health.Dicom.Web/Program.cs
@@ -32,7 +32,7 @@ namespace Microsoft.Health.Dicom.Web
 
                     builder.AddDevelopmentAuthEnvironmentIfConfigured(builtConfig, "DicomServer");
                 })
-                .ConfigureKestrel(option => option.Limits.MaxRequestBodySize = long.MaxValue)
+                .ConfigureKestrel(option => option.Limits.MaxRequestBodySize = uint.MaxValue)
                 .UseStartup<Startup>()
                 .Build();
